requirements.txt
python-dotenv
aiohttp
openai
